refactor(c3): move enums to const
