+++ midealocal/devices/c3/const.py
@@ -1,0 +1,64 @@
+"""Midea local C3 device const."""
+
+from enum import IntEnum, StrEnum
+
+
+class DeviceAttributes(StrEnum):
+    """Midea C3 device attributes."""
+
+    zone1_power = "zone1_power"
+    zone2_power = "zone2_power"
+    dhw_power = "dhw_power"
+    zone1_curve = "zone1_curve"
+    zone2_curve = "zone2_curve"
+    disinfect = "disinfect"
+    fast_dhw = "fast_dhw"
+    zone_temp_type = "zone_temp_type"
+    zone1_room_temp_mode = "zone1_room_temp_mode"
+    zone2_room_temp_mode = "zone2_room_temp_mode"
+    zone1_water_temp_mode = "zone1_water_temp_mode"
+    zone2_water_temp_mode = "zone2_water_temp_mode"
+    mode = "mode"
+    mode_auto = "mode_auto"
+    zone_target_temp = "zone_target_temp"
+    dhw_target_temp = "dhw_target_temp"
+    room_target_temp = "room_target_temp"
+    zone_heating_temp_max = "zone_heating_temp_max"
+    zone_heating_temp_min = "zone_heating_temp_min"
+    zone_cooling_temp_max = "zone_cooling_temp_max"
+    zone_cooling_temp_min = "zone_cooling_temp_min"
+    tank_actual_temperature = "tank_actual_temperature"
+    room_temp_max = "room_temp_max"
+    room_temp_min = "room_temp_min"
+    dhw_temp_max = "dhw_temp_max"
+    dhw_temp_min = "dhw_temp_min"
+    target_temperature = "target_temperature"
+    temperature_max = "temperature_max"
+    temperature_min = "temperature_min"
+    status_heating = "status_heating"
+    status_dhw = "status_dhw"
+    status_tbh = "status_tbh"
+    status_ibh = "status_ibh"
+    total_energy_consumption = "total_energy_consumption"
+    total_produced_energy = "total_produced_energy"
+    outdoor_temperature = "outdoor_temperature"
+    silent_mode = "silent_mode"
+    SILENT_LEVEL = "silent_level"
+    eco_mode = "eco_mode"
+    tbh = "tbh"
+    error_code = "error_code"
+
+
+class C3SilentLevel(IntEnum):
+    """C3 Silent Level."""
+
+    OFF = 0x0
+    SILENT = 0x1
+    SUPER_SILENT = 0x3
+
+
+class C3DeviceMode(IntEnum):
+    """C3 Device Mode."""
+
+    COOL = 2
+    HEAT = 3
